Add base functions(car and order)

+++ src/main/java/com/example/car/model/CarEntity.java
@@ -42,5 +42,5 @@ public class CarEntity {
     /**
      *
      */
-    private Byte deleteFlag;
+    private Integer deleteFlag;
 }
