+++ src/main/java/com/example/car/model/UserEntity.java
@@ -32,5 +32,5 @@ public class UserEntity {
     /**
      *
      */
-    private Byte deleteFlag;
+    private Integer deleteFlag;
 }
